refactor : IssueList 역할 분리
